move "ZenFSMetrics" to file "zbd_stat.h"
